fix: E2E quality audit remediation — 80 findings across 6 phases (#630)
* fix(sync): preserve active cycle during merge, track backup session IDs inline

[14] mergePortalCycles + mergeAllPullData: use existing.is_active instead
of hardcoding 0L — prevents silent deactivation when portal sends no
active cycle. Regression test added with matching cycle ID.

[17] DataBackupManager.importFromJson: build importedSessionIds from
actual successful insertions (mutableSetOf inline) instead of
recomputing from the manifest. Prevents FK crashes that rolled back
the entire import transaction.


* fix(safety): clamp VBT 1RM estimate at per-cable hardware max, require 3+ loads

[15] AssessmentEngine: add .coerceAtMost(110.0) upper bound on estimated
1RM (MAX_WEIGHT_PER_CABLE_KG). Prevents extrapolated values far above
the machine physical limit from propagating as routine scaling baselines.

VelocityOneRepMaxEstimator: raise MIN_DISTINCT_LOADS from 2 to 3.
Two-point OLS always yields r²=1.0 (vacuous quality gate); requiring 3+
points ensures the regression quality is meaningful. Tests updated.


* fix(ui): synchronize volume chart scroll state across canvas and labels

[16] Share single rememberScrollState() across all three horizontalScroll
modifiers in VolumeTrendChart so the bar canvas, bar labels, and line
labels scroll in lockstep.

Audit finding: [16]


* fix(sync): harden mobile sync — retry guards, DTO cleanup, dedup merge helper

- SyncModels: add totalTimeSeconds to GamificationStatsSyncDto, delete
  legacy SyncPushRequest/SyncPullResponse wrappers [53][54]
- PortalPullAdapter: wire totalTimeSeconds through pull merge path [53]
- SyncTriggerManager: reinstate MAX_CONSECUTIVE_FAILURES persistent error
  guard, move throttle read inside platform lock for atomicity [55][78]
- PortalApiClient: double-check-after-lock in forceRefresh to prevent
  redundant token refreshes under contention [77]
- MigrationManager: gate one-time repair functions behind persisted
  version flags via Settings to prevent repeated full-table scans [51]
- DomainModule: pass Settings to MigrationManager for flag persistence [51]
- SqlDelightSyncRepository: extract 130-line duplicate routine-exercise
  merge block into mergePortalExercisesForRoutine helper, add missing
  PullRoutineExerciseDto import [52]


* fix(ui): Phase 5 mobile app-layer mediums + CI/config — 25 audit findings

BLE: true exponential backoff [47], disconnect before retry [49],
cancel stateObserverJob on exhaustion [48].
Domain: @Volatile firstRepMcv + guard mcv>0 baseline [56][57].
Charts: CircleChart full-sweep accumulation [59], VolumeTrendChart
Animatable replay [60].
UI: GroupedExerciseList isLoading param [63], SwipeableExerciseRow
pointerInput key [64], ProgressionSuggestion semantic accent colors
[65] + unit-aware chip labels [66].
Components: dead AnimatedRepCounter param [58], dead WorkoutTab
position-bar block [71], MotionStartDetector DROP_OLDEST buffer [68].
Screens: remove isAutoConnecting collection from 7 screens [69].
VM: WorkoutHud latch → SideEffect [70], saveCycle entry guard [72],
ExerciseConfigViewModel standard generateUUID [73].
iOS: CsvExporter writeToFile check [74], HealthPermissionSettings
open via UIApplicationOpenSettingsURLString [75].
CI: libs.versions.toml in all 4 cache keys [12], release versionCode
guard via tasks.configureEach [13].


* refactor: remove dead code — Greeting.kt, ActiveWorkoutCard, unused StateFlows, empty inits

Phase 6 cleanliness sweep: delete KMP boilerplate Greeting.kt, dead ActiveWorkoutCard
composable, unused personalBests StateFlow chain (HistoryManager→MainViewModel),
dead isRefreshing variable, unreachable TUTBeast when branch (mapped away at line 114),
unused EulaViewModel methods, empty init{} block, dead null-coalesce, stale placeholder
comments, displaced velocity zone comments in BiomechanicsModels.


* fix(ble): protocol parser tick overflow, packet capture sync, fault decoder dedup

Phase 6 BLE cluster: use Long for tick counters to prevent Int overflow after ~24 days,
add synchronized{} to BlePacketCapture buffer access, deduplicate fault decoder messages,
fix displaced HandleStateDetector comments, correct ECHO packet size comment (29→32 bytes),
rename MAX_WEIGHT_KG→MAX_TOTAL_WEIGHT_KG in BleConstants for clarity.


* fix(sync): duration overflow, LWW merge defaults, API double-encode, DTO type fixes

Phase 6 sync-data cluster: use Long for PortalSyncAdapter duration to prevent Int
overflow on long sessions, fix SqlDelightSyncRepository mergeSessionsLww timestamp
default and GamificationStats Float→Long truncation, fix mergePortalCycles defaults,
prevent PortalApiClient double URL-encoding, update SyncModels duration type to Long.


* fix(ui): composition side-effects, constant consolidation, DI registration, config cleanup

Phase 6 UI/platform/config cluster: fix WeightChangePerRepControl composition side-effect,
correct BulkWeightAdjustDialog upper bound annotation, replace hardcoded color in
SafetyEventsCard with AccessibilityTheme, fix iOS ConnectivityChecker initial state,
fix Android DeviceInfo hardcoded version, consolidate KG_TO_LB constant, register
ExerciseConfigViewModel in Koin DI, fix CycleEditorViewModel non-atomic initialize(),
remove JitPack repository, add .gitignore entries.


* fix: address PR review findings — threading, constants, Compose patterns

- BlePacketCapture: replace runBlocking+coroutine Mutex with synchronized()
  to avoid deadlock risk in non-suspend BLE callback hot path
- MonitorDataProcessor: validate per-cable load against MAX_WEIGHT_PER_CABLE_KG
  (110kg) instead of MAX_TOTAL_WEIGHT_KG (220kg) — was a no-op
- BiomechanicsEngine: replace println with Kermit Logger.w for proper log
  level filtering and release-build stripping
- CycleEditorViewModel: reset shouldProceed=false in CAS retry if-block to
  prevent race where lambda re-execution leaks a stale true
- KableBleConnectionManager: rethrow CancellationException in disconnect
  catch blocks during connection retry cleanup (×2 sites)
- AssessmentEngine: use Constants.MAX_WEIGHT_PER_CABLE_KG instead of
  hardcoded 110.0 for 1RM clamp ceiling
- WorkoutHud: replace SideEffect+mutableState latch with guarded
  composition-time update to avoid recomposition loop risk
- SwipeableExerciseRow: use pointerInput(Unit)+rememberUpdatedState instead
  of pointerInput(isRevealed) to avoid cancelling active drag gestures


* fix(test): MonitorDataProcessorTest ticks Int→Long to match MonitorPacket

Phase 6 changed MonitorPacket.ticks from Int to Long for tick overflow
prevention but missed updating the test helper's parameter type, causing
compileAndroidHostTest to fail on CI.


* fix(test): update tests for Phase 6 fault labels and SideEffect refactor

- DiagnosticFaultDecoderTest: update expected labels to match Phase 6's
  "Message failure (RX)" / "(TX)" disambiguation (was "Message failure")
- RestTimerProgressionWiringTest: expect SideEffect pattern instead of
  LaunchedEffect after Phase 6 composition side-effect refactor


* fix(kmp): replace JVM-only synchronized/Volatile with KMP-compatible alternatives

- BlePacketCapture: replace `synchronized(packets)` with atomicfu
  `reentrantLock().withLock` — synchronized is JVM-only, breaks iOS compile
- BlePacketCapture: move logDualInterpretation call outside lock to
  avoid holding lock during I/O
- BiomechanicsEngine: add missing `import kotlin.concurrent.Volatile` —
  without it @Volatile resolves to kotlin.jvm.Volatile, unresolved on Native
- Add atomicfu as explicit shared module dependency (already in version
  catalog, used by kable in this repo)


* Separate period selector scroll state and relax sync retries

+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/data/ble/ProtocolModels.kt
@@ -6,7 +6,7 @@ package com.devil.phoenixproject.data.ble
  * Created by parseMonitorPacket().
  */
 data class MonitorPacket(
-    val ticks: Int,
+    val ticks: Long,
     val posA: Float, // mm
     val posB: Float, // mm
     val loadA: Float, // kg
